Finish example C++ code
